Move files into a new namespace
