--- Resources/config/experimental.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" ?>
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <!--
-        @internal this file should not be referenced inside your project
-
-        NOTE: this are the service definitions when using the "experimental" storage
-        The service definitions when using the "phpcr" storage can
-        be found in the "services.xml" file in this directory.
-    -->
-
-    <services>
-        <!-- Message Handler -->
-        <service id="sulu_article.create_article_handler" class="Sulu\Bundle\ArticleBundle\Application\MessageHandler\CreateArticleMessageHandler">
-            <argument type="service" id="sulu_article.article_repository"/>
-            <argument type="tagged_iterator" tag="sulu_commerce.article_mapper"/>
-
-            <tag name="messenger.message_handler"/>
-        </service>
-
-        <service id="sulu_article.modify_article_handler" class="Sulu\Bundle\ArticleBundle\Application\MessageHandler\ModifyArticleMessageHandler">
-            <argument type="service" id="sulu_article.article_repository"/>
-            <argument type="tagged_iterator" tag="sulu_commerce.article_mapper"/>
-
-            <tag name="messenger.message_handler"/>
-        </service>
-
-        <service id="sulu_article.remove_article_handler" class="Sulu\Bundle\ArticleBundle\Application\MessageHandler\RemoveArticleMessageHandler">
-            <argument type="service" id="sulu_article.article_repository"/>
-
-            <tag name="messenger.message_handler"/>
-        </service>
-
-        <service id="sulu_article.apply_workflow_transition_article_handler" class="Sulu\Bundle\ArticleBundle\Application\MessageHandler\ApplyWorkflowTransitionArticleMessageHandler">
-            <argument type="service" id="sulu_article.article_repository"/>
-            <argument type="service" id="sulu_content.content_workflow"/>
-
-            <tag name="messenger.message_handler"/>
-        </service>
-
-        <service id="sulu_article.copy_locale_article_handler" class="Sulu\Bundle\ArticleBundle\Application\MessageHandler\CopyLocaleArticleMessageHandler">
-            <argument type="service" id="sulu_article.article_repository"/>
-            <argument type="service" id="sulu_content.content_copier"/>
-
-            <tag name="messenger.message_handler"/>
-        </service>
-
-        <!-- Mapper -->
-        <service id="sulu_article.article_content_mapper" class="Sulu\Bundle\ArticleBundle\Application\Mapper\ArticleContentMapper">
-            <argument type="service" id="sulu_content.content_persister"/>
-
-            <tag name="sulu_commerce.article_mapper"/>
-        </service>
-
-        <!-- Sulu Integration -->
-        <service id="sulu_article.article_admin" class="Sulu\Bundle\ArticleBundle\Infrastructure\Sulu\Admin\ArticleAdmin">
-            <!-- @internal this service is internal and should not be used by your project -->
-
-            <argument type="service" id="sulu_admin.view_builder_factory" />
-            <argument type="service" id="sulu_content.content_view_builder_factory" />
-            <argument type="service" id="sulu_security.security_checker" />
-            <argument type="service" id="sulu.core.localization_manager" />
-
-            <tag name="sulu.context" context="admin"/>
-            <tag name="sulu.admin" />
-        </service>
-
-        <!-- Repositories -->
-        <service id="sulu_article.article_repository" class="Sulu\Bundle\ArticleBundle\Infrastructure\Doctrine\Repository\ArticleRepository">
-            <argument type="service" id="doctrine.orm.entity_manager" />
-            <argument type="service" id="sulu_content.dimension_content_query_enhancer"/>
-        </service>
-
-        <service id="Sulu\Bundle\ArticleBundle\Domain\Repository\ArticleRepositoryInterface" class="sulu_article.article_repository" />
-
-        <!-- Controllers -->
-        <service id="sulu_article.admin_article_controller" class="Sulu\Bundle\ArticleBundle\UserInterface\Controller\Admin\ArticleController" public="true">
-            <argument type="service" id="sulu_article.article_repository" />
-            <argument type="service" id="sulu_article.message_bus"/>
-            <argument type="service" id="fos_rest.view_handler" />
-            <!-- TODO remove additional services when not longer needed -->
-            <argument type="service" id="sulu_content.content_manager"/>
-            <argument type="service" id="sulu_core.list_builder.field_descriptor_factory"/>
-            <argument type="service" id="sulu_core.doctrine_list_builder_factory"/>
-            <argument type="service" id="sulu_core.doctrine_rest_helper"/>
-
-            <tag name="sulu.context" context="admin"/>
-        </service>
-
-        <!-- Preview -->
-        <service id="sulu_article.article_preview_provider" class="Sulu\Bundle\ContentBundle\Content\Infrastructure\Sulu\Preview\ContentObjectProvider">
-            <argument type="service" id="doctrine.orm.entity_manager" />
-            <argument type="service" id="sulu_content.content_resolver"/>
-            <argument type="service" id="sulu_content.content_data_mapper" />
-            <argument>%sulu.model.article.class%</argument>
-            <argument type="constant">Sulu\Bundle\ArticleBundle\Infrastructure\Sulu\Admin\ArticleAdmin::SECURITY_CONTEXT</argument>
-
-            <tag name="sulu.context" context="admin"/>
-            <tag name="sulu_preview.object_provider" provider-key="articles"/>
-        </service>
-    </services>
-</container>
