--- Resources/config/routing_experimental_admin_api.yaml
@@ -1,12 +1,0 @@
-sulu_article.admin_article_api:
-    type: rest
-    resource: sulu_article.admin_article_controller
-    name_prefix: sulu_article.
-
-
-sulu_article.post_article_trigger:
-    path: /articles/{id}.{_format}
-    methods: POST
-    defaults:
-        _controller: sulu_article.admin_article_controller::postTriggerAction
-        _format: json
