+++ src/Application/Message/RemoveArticleMessage.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) Sulu GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Article\Application\Message;
+
+/**
+ * @experimental
+ */
+class RemoveArticleMessage
+{
+    /**
+     * @var array{
+     *     uuid?: string
+     * }
+     */
+    private $identifier;
+
+    /**
+     * @param array{
+     *     uuid?: string
+     * } $identifier
+     */
+    public function __construct(array $identifier)
+    {
+        $this->identifier = $identifier;
+    }
+
+    /**
+     * @return array{
+     *     uuid?: string
+     * }
+     */
+    public function getIdentifier(): array
+    {
+        return $this->identifier;
+    }
+}
